Clean out and rename location file
